# Conflicts:
#	frontend/node_modules/.vite/deps/_metadata.json
#	frontend/node_modules/.vite/deps/antd.js
#	frontend/node_modules/.vite/deps/antd.js.map
#	frontend/node_modules/.vite/deps/antd_locale_zh_CN.js.map
#	frontend/node_modules/.vite/deps/axios.js.map
#	frontend/node_modules/.vite/deps/chunk-GF7PSAJX.js
#	frontend/node_modules/.vite/deps/chunk-YLZ34CCM.js
#	frontend/node_modules/.vite/deps/chunk-ZNKPWGXJ.js
#	frontend/node_modules/.vite/deps/react-dom.js
#	frontend/node_modules/.vite/deps/react-dom_client.js
#	frontend/node_modules/.vite/deps/react-router-dom.js
#	frontend/node_modules/.vite/deps/react-router-dom.js.map
#	frontend/node_modules/.vite/deps/react.js
#	frontend/node_modules/.vite/deps/react_jsx-dev-runtime.js
#	frontend/node_modules/.vite/deps/react_jsx-dev-runtime.js.map
#	frontend/node_modules/.vite/deps/react_jsx-runtime.js
#	frontend/node_modules/.vite/deps/react_jsx-runtime.js.map

+++ backend/src/main/java/com/tree/plms/model/dto/response/DailyStatsDTO.java
@@ -16,10 +16,10 @@ public class DailyStatsDTO {
     private String date;
     
     @Schema(description = "入场车辆数", example = "100")
-    private Integer entryCount;
-    
+    private Long entryCount;
+
     @Schema(description = "出场车辆数", example = "95")
-    private Integer exitCount;
+    private Long exitCount;
     
     @Schema(description = "当日收入", example = "5000.00")
     private BigDecimal totalRevenue;
